moved parsing last and firstname to abstract class
