+++ member-service/src/main/java/com/hedvig/memberservice/events/ZignSecSuccessfulAuthenticationEvent.kt
@@ -4,9 +4,9 @@ package com.hedvig.memberservice.events
 data class ZignSecSuccessfulAuthenticationEvent(
     val memberId: Long,
     val ssn: String,
-    val providerJsonResponse: String,
+    override val providerJsonResponse: String,
     val authenticationMethod: AuthenticationMethod
-) {
+): ZignSecIdentityJson() {
     enum class AuthenticationMethod {
         NORWEGIAN_BANK_ID,
         DANISH_BANK_ID
